[UPDATE] update type to be float64 to use with flutter

+++ endpoints/review/review_addReview.go
@@ -64,7 +64,7 @@ func AddDormReview(c *fiber.Ctx) error {
 	}
 
 	//แก้ชั่วคราว เพราะเปลี่ยน overall -> float
-	overall := float32(*body.RatingOverall)
+	overall := float64(*body.RatingOverall)
 
 	DormReview := &model.Review{
 		UserId:         body.UserId,
