+++ loaders/mysql/model/reviews.go
@@ -9,7 +9,7 @@ type Review struct {
 	UserId         *string    `json:"user_id" gorm:"not null;primaryKey;autoIncrement:false;"`
 	User           *User      `json:"user" gorm:"foreignKey:UserId;references:Id;not null"`
 	Review         *string    `json:"review" gorm:"not null"`
-	RatingOverall  *float32   `json:"rating_overall" gorm:"not null"`
+	RatingOverall  *float64   `json:"rating_overall" gorm:"not null"`
 	RatingPrice    *int       `json:"price" gorm:"not null"`
 	RatingLocation *int       `json:"location" gorm:"not null"`
 	RatingFacility *int       `json:"facility" gorm:"not null"`
